# openapi/openapi.yaml
openapi: 3.1.0

info:
  title: Travelog API
  version: 0.1.0
  description: |
    REST API for Travelog MVP1.

    Functional behavior is defined by:
    `doc/functional-requirements-mvp1.md`

    Technical architecture is defined by:
    `doc/technical-design-mvp1.md`

servers:
  - url: /api
    description: Travelog API

tags:
  - name: Health
  - name: Scans
  - name: Trips
  - name: Trip Operations
  - name: Administrative Areas
  - name: Exclusion Zones
  - name: Settings

paths:
  /health:
    get:
      tags: [Health]
      operationId: getHealth
      summary: Check API health
      responses:
        "200":
          description: API is available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /scans:
    post:
      tags: [Scans]
      operationId: startScan
      summary: Start a photo scan
      description: |
        Starts an asynchronous scan of the selected directory.
        The directory must be relative to the configured photo root.

        Only one scan may run at a time.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartScanRequest"
      responses:
        "202":
          description: Scan started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "400":
          description: Invalid scan request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Another scan is already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    get:
      tags: [Scans]
      operationId: listScans
      summary: List scan history
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Scan history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanList"

  /scans/{scanId}:
    get:
      tags: [Scans]
      operationId: getScan
      summary: Get scan status and progress
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Scan status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "404":
          description: Scan not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /trips:
    get:
      tags: [Trips]
      operationId: listTrips
      summary: List trips
      description: |
        By default only active trips are returned.
        Archived/superseded trips can be requested explicitly.

        Results are ordered by start date descending by default.
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/TripListStatus"
          description: Filter by trip status.
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: |
            Free-text search by trip name or year/month.
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - startDateDesc
              - startDateAsc
          default: startDateDesc
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Trip list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripList"

    post:
      tags: [Trips]
      operationId: createTrip
      summary: Create a trip manually
      description: |
        Creates a trip from an explicit manual operation.

        This endpoint is not used by automatic scan processing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Invalid trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Trip overlaps another active trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /trips/{tripId}:
    get:
      tags: [Trips]
      operationId: getTrip
      summary: Get trip details
      parameters:
        - $ref: "#/components/parameters/TripId"
      responses:
        "200":
          description: Trip details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    patch:
      tags: [Trips]
      operationId: updateTrip
      summary: Modify a trip manually
      description: |
        Allows the manual modifications supported by MVP1:
        rename and/or change start/end dates.

        Automatic processing never modifies an existing trip.

        Active trips must not overlap temporally.
      parameters:
        - $ref: "#/components/parameters/TripId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTripRequest"
      responses:
        "200":
          description: Updated trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "400":
          description: Invalid modification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Modification would overlap another active trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /trips/{tripId}/split:
    post:
      tags: [Trips]
      operationId: splitTrip
      summary: Split a trip
      description: |
        Splits a trip at the specified date.

        The split date belongs to the second resulting trip.
        The original trip remains in history.
      parameters:
        - $ref: "#/components/parameters/TripId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitTripRequest"
      responses:
        "200":
          description: Split result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripOperationResult"
        "400":
          description: Invalid split date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Trip not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Split would violate trip constraints
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /trips/merge:
    post:
      tags: [Trips]
      operationId: mergeTrips
      summary: Merge two or more trips
      description: |
        Merges two or more trips into a new trip.

        Original trips are retained in history and marked as superseded.
        The resulting trip keeps a relationship to the source trips.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeTripsRequest"
      responses:
        "200":
          description: Merge result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripOperationResult"
        "400":
          description: Invalid merge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: One or more trips were not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Trips cannot be merged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /settings:
    get:
      tags: [Settings]
      operationId: getSettings
      summary: Get current settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

    put:
      tags: [Settings]
      operationId: updateSettings
      summary: Update settings
      description: |
        Updates one or more setting values.

        Only provided fields are updated; omitted fields remain unchanged.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsRequest"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "400":
          description: Invalid settings update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    post:
      tags: [Settings]
      operationId: recalculate
      summary: Trigger trip recalculation
      description: |
        Requests an explicit trip recalculation.

        The operation returns immediately with ACCEPTED status.
        Actual recalculation proceeds asynchronously.
      responses:
        "202":
          description: Recalculation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recalculation"
        "400":
          description: Invalid recalculation request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /exclusion-zones:
    get:
      tags: [Exclusion Zones]
      operationId: listExclusionZones
      summary: List exclusion zones
      responses:
        "200":
          description: Exclusion zone list
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExclusionZone"

    post:
      tags: [Exclusion Zones]
      operationId: createExclusionZone
      summary: Create an exclusion zone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExclusionZoneRequest"
      responses:
        "201":
          description: Exclusion zone created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExclusionZone"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Administrative area not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    delete:
      tags: [Exclusion Zones]
      operationId: deleteExclusionZone
      summary: Delete an exclusion zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "204":
          description: Deleted
        "404":
          description: Exclusion zone not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /admin-areas/search:
    get:
      tags: [Administrative Areas]
      operationId: searchAdministrativeAreas
      summary: Search administrative areas
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: Search query
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Maximum results to return
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdministrativeArea"
        "400":
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"


components:
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-based)
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Page size
    ScanId:
      name: scanId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    TripId:
      name: tripId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1

  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          example: TRIP_NOT_FOUND
        message:
          type: string
          example: Trip not found
        details:
          type: object
          additionalProperties: true

    HealthResponse:
      type: object
      additionalProperties: false
      required: [status, timestamp, uptime_seconds, version, environment]
      properties:
        status:
          type: string
          enum: [ok]
        timestamp:
          type: string
          format: date-time
        uptime_seconds:
          type: number
        version:
          type: string
          example: "0.1.0"
        environment:
          type: string
          example: development

    Trip:
      type: object
      additionalProperties: false
      required: [id, name, startDate, endDate, autoGenerated, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        autoGenerated:
          type: boolean
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
          format: date-time

    CreateTripRequest:
      type: object
      additionalProperties: false
      required: [startDate, endDate]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    UpdateTripRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date

    SplitTripRequest:
      type: object
      additionalProperties: false
      required: [splitDate]
      properties:
        splitDate:
          type: string
          format: date

    MergeTripsRequest:
      type: object
      additionalProperties: false
      required: [tripIds]
      properties:
        tripIds:
          type: array
          minItems: 2
          uniqueItems: true
          items:
            type: integer
            format: int64
            minimum: 1
        title:
          type: [string, null]
          minLength: 1

    TripOperationType:
      type: string
      enum: [SPLIT, MERGE]

    TripOperation:
      type: object
      additionalProperties: false
      required: [id, type, createdAt, sourceTripIds, resultingTripIds]
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: "#/components/schemas/TripOperationType"
        createdAt:
          type: string
          format: date-time
        sourceTripIds:
          type: array
          items:
            type: integer
            format: int64
        resultingTripIds:
          type: array
          items:
            type: integer
            format: int64

    TripOperationResult:
      type: object
      additionalProperties: false
      required: [operation, trips]
      properties:
        operation:
          $ref: "#/components/schemas/TripOperation"
        trips:
          type: array
          items:
            $ref: "#/components/schemas/Trip"

    TripList:
      type: object
      additionalProperties: false
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Trip"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer

    Settings:
      type: object
      additionalProperties: false
      required: [minimumPhotosPerVisit, consecutiveDaysWithoutPhotosBeforeClosing]
      properties:
        minimumPhotosPerVisit:
          type: integer
          minimum: 1
          default: 1
        consecutiveDaysWithoutPhotosBeforeClosing:
          type: integer
          minimum: 0
          default: 3

    UpdateSettingsRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        minimumPhotosPerVisit:
          type: integer
          minimum: 1
        consecutiveDaysWithoutPhotosBeforeClosing:
          type: integer
          minimum: 0

    Recalculation:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [ACCEPTED]

    ExclusionZone:
      type: object
      additionalProperties: false
      required: [id, administrativeArea]
      properties:
        id:
          type: integer
          format: int64
        administrativeArea:
          $ref: "#/components/schemas/AdministrativeArea"

    CreateExclusionZoneRequest:
      type: object
      additionalProperties: false
      required: [administrativeAreaId]
      properties:
        administrativeAreaId:
          type: integer
          format: int64
          minimum: 1

    AdministrativeArea:
      type: object
      additionalProperties: false
      required: [id, datasetSource, countryCode, name, adminLevel]
      properties:
        id:
          type: integer
          format: int64
        datasetSource:
          type: string
        countryCode:
          type: string
        name:
          type: string
        adminLevel:
          type: integer
        parentId:
          type: integer
          format: int64

    TripListStatus:
      type: string
      enum: [active, archived]